Deliverable 2 of Iteration 1 is complete: Unittests have been written for all public methods of passenger class, constructors, and boolean functions of stop, route, and bus classes.
Test: Google Unittesting is complete for
passenger_UT.cc
stop_UT.cc
route_UT.cc
bus_UT.cc

However, an explanation is needed for Bus::Move(). There is an arithmetic
error in this method that causes it to not apply the correct distance between
certain stops. I believe the logic of this method is that it should return
true when the bus does move from or to a stop, return false when at the first
stop of the route, and return false when both routes are complete. This was
my expectation for the Bus::Move() unit testing. Yet, Bus::Move() only
returns true when distance_remaining_ is greater than 1 at the conclusion
of the call to Route::GetNextStopDistance(). So, it returns false far too
often which is why the unit test for false passes and the unit test for true
fails.