feat(bus.cc and config.txt): prints information to terminal when a bus skips a stop
In terms of our simulation it is not necessary for a bus to "literally" skip a
stop. It need only check that it unloaded or loaded passengers when at the stop.
This feature prints to the terminal the bus id and stop id when that bus skips a stop.
This feature appears in the Bus::Move() function.

To better test this feature the probability a stop generates a passenger for all stops
in route 2 was set to 0. Hence the modification to config.txt.