feat(bus, PassengerUnloader,  VisualizationSimulator, new class FileWriter): #16 Singleton file logger
This encompases lab 16 and Priority level 2 from Iteration 3.
Using the singleton pattern a filewriter class was made. It writes to one of two files based on whether the data is from a bus or passenger object. The data it writes is from either Bus::Report or Passenger::Report. To incorporate it into the program, changes were made to visualizationSimulator::Update and PassengerUnloader::UnloadPassengers. Additionally, the bus now keeps track of the total number of passengers that have ridden the bus and ouputs this data in its Bus::Report method.