feat(IBusDecorator, BusColorDecorator, IBus, Bus, and the new member of the family BusIntensityDecorator): #18 Implemented an alpha channel intensity decorator
To appropriately implement this new decorator I had to refactor the existing code for BusColorDecorator::SetColor method so that neither decorator will step on each other's toes.

Previously, SetColor used a color object to change color values to either maroon or gold. But this included the alpha values. If left as is, these values would overwrite those set by BusIntensityDecorator.

Now, SetColor retrieves the bus_data_ object from the bus it's wrapping, change the RGB color values, then uses the new SetBusData method to send it back to the bus. BusIntensityDecorator::SetIntensity works on the same principle but for setting the alpha channel based on how many passengers are on the bus.

Lastly, there is an error that a new issue will be created to address. I'm not able to move the method definitions from the header files of BusColorDecorator and BusIntensityDecorator to a .cc file without Bus losing the ability to inherit from IBus.