feat(template observer pattern): #19 Save progress on implementing the template observer
I have encountered significant issues working on the template observer where I am unable to test the code changes I am making. And it is made worse by the many code changes I have made
I plan to go back to the drawing board and work on it in a more piecemeal fashion.