fix(BusWebObserver): #20 Bus Observer Pattern is functional again
I had made this issue/branch from the working devel branch solely to fix the broken Bus Observer design. Since adding the bus decorator the busdata was no longer being displayed. The solution was to add pass-through methods in the decorators for RegisterObserver and ClearObservers and to add the virtual keyword to the method declarations in IObservable.h.