feat(Stop Observer): #21 Stop observer is complete along with the template pattern
I pulled an audible in the middle of this branch. I had initially intended to write the stop observer before changing the observer pattern to templates. But after I got the Bus Observer to work after changing the code to specificially refer to the bus, I decided to get the template to work for the bus observer. Progress was quick and in almost no time the bus observer was functioning with the template base class. After that it was smooth sailing implementing the stop observer.

All major priority levels of Iteration 3 are complete.